feat: 新增多项系统功能并完善兼容性

+++ includes/kernel/syscall/linux_compat.h
@@ -8,6 +8,6 @@
 
 #include "kernel/syscall/linux_abi.h"
 
-uint32_t linux_compat_handler(struct Registers *r);
+int64_t linux_compat_handler(struct Registers *r);
 
 #endif
